requirements.txt
Django==1.8.4
django-extensions==1.5.5
factory-boy==2.5.2
ipdb==0.8.1
ipython==4.0.0
MySQL-python==1.2.5
pytz==2015.4
sphinx-autobuild==0.5.2
Sphinx==1.3.1
djangorestframework==3.2.3
django-annoying==0.8.4
